ISSUE-10:(feat) corrections after the review + corrected some styles
