+++ src/components/filter.tsx
@@ -1,7 +1,9 @@
 export default function Filter() {
     return (
         <div className="w-80 bg-gray-800 p-4">
-            <div className="text-white font-bold mb-4">Фильтры</div>
+            <div className="text-white font-bold mb-4">
+                <p>Фильтры</p>
+                </div>
             {/* Контент для фильтров */}
         </div>
     );
